respaldo/requirements.txt
flask
gunicorn
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
requests
boto3
python-dotenv
flask-cors
firebase-admin
hashlib2